luna-service=145, cmake-modules-webos=12
Updated luna-service and luna-prefs to build with cmake-modules-webos.

Updated service-bus script to find luna-service binaries in /usr/...

Luna-sysmgr, Luna-prefs, and Activitymanager are still looking in the
wrong place for luna-service headers and some components are linking
against liblunaservice.so instead of libluna-service2 and/or not yet
looking for luna-service in staging/usr/lib instead of staging/lib;
will resolve separately since other work is waiting for this update
to the latest (cmake-built) luna-service2.

Open-webOS-DCO-1.0-Signed-off-by: Steve Lemke <steve.lemke@palm.com>

Change-Id: I168e88c64e45074fed51d7175457dfaaf5bbe4b7
Reviewed-on: http://g2g.palm.com/1073
Reviewed-by: Build Verification
Reviewed-by: Steve Lemke <steve.lemke@palm.com>
Tested-by: Steve Lemke <steve.lemke@palm.com>
Reviewed-by: Eric Blade <eric.blade@palm.com>
Reviewed-by: Keith Derrick <keith.derrick@palm.com>