Restructure method of customizing the desktop build
Significant structural change to the build-webos-desktop script to
facilitate use of cloned copies of components.  This recognizes the
primary development role of the desktop build.

The user can now specify an alternate location for the desktop binaries
instead of the traditional location of $HOME/luna-desktop-binaries by
using the local.sh file in the customize folder.  A default version of
this file is provided and will be exported to the customize folder when
the desktop build script is first run.  Alternatively the user can
manually create the customize folder, copy the local.sh.default file
into it as local.sh, and then edit it to set the location.

Previously users had to modify the build-webos-desktop script to have
the script build from a cloned copy of a component.  The location of a
clone component can instead be added to the locations.sh file in the
customize folder.  A default version of this file is provided and will
be exported to the customize folder when the desktop build script is
first run.
To have the script build from a cloned copy of (say) webappmanager,
uncomment the WEBAPPMANAGER_DIR line in locations.sh and replace the
default location with the location of your cloned copy of the component.
This ensures that changes you make aren't overwritten by the build
script, which currently happens when a user modifies the versions
downloaded by the script.

Open-webOS-DCO-1.0-Signed-off-by: Roger Stringer <roger.stringer@palm.com>

Change-Id: Iad33d1ac58d5cd6e3109e0d816f217e1c4bf4600
Reviewed-on: http://g2g.palm.com/1308
Reviewed-by: Build Verification
Reviewed-by: Roger Stringer <roger.stringer@palm.com>
Tested-by: Roger Stringer <roger.stringer@palm.com>
Reviewed-by: Steve Lemke <steve.lemke@palm.com>